+++ sdk/confidentialledger/Azure.Security.CodeTransparency/tests/CcfReceiptVerifierTests.cs
@@ -67,9 +67,9 @@ namespace Azure.Security.CodeTransparency.Tests
             byte[] receiptBytes = readFileBytes("receipt.cose");
             byte[] inputSignedPayloadBytes = readFileBytes("input_signed_claims");
 
-            Response<JwksDocument> key = client.GetPublicKeys();
+            Response<CodeTransparencyVerificationKeySet> key = client.GetPublicKeys();
 
-            var exception = Assert.Throws<InvalidOperationException>(() => CcfReceiptVerifier.VerifyTransparentStatementReceipt(key.Value.Keys[0], receiptBytes, inputSignedPayloadBytes));
+            var exception = Assert.Throws<InvalidOperationException>(() => CcfReceiptVerifier.Verify(receiptBytes, inputSignedPayloadBytes, key.Value.Keys[0]));
             StringAssert.Contains(expected: "KID mismatch", exception.Message);
 #endif
         }
@@ -100,9 +100,9 @@ namespace Azure.Security.CodeTransparency.Tests
             byte[] receiptBytes = readFileBytes("receipt.cose");
             byte[] inputSignedPayloadBytes = readFileBytes("input_signed_claims");
 
-            Response<JwksDocument> key = client.GetPublicKeys();
+            Response<CodeTransparencyVerificationKeySet> key = client.GetPublicKeys();
 
-            var exception = Assert.Throws<InvalidOperationException>(() => CcfReceiptVerifier.VerifyTransparentStatementReceipt(key.Value.Keys[0], receiptBytes, inputSignedPayloadBytes));
+            var exception = Assert.Throws<InvalidOperationException>(() => CcfReceiptVerifier.Verify(receiptBytes, inputSignedPayloadBytes, key.Value.Keys[0]));
             StringAssert.Contains(expected: "Claim digest mismatch", exception.Message);
 #endif
         }