+++ sdk/storage/Azure.Storage.Blobs.Batch/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 12.27.0-beta.2 (Unreleased)
+
+### Bugs Fixed
+- Fixed bug where header values were not being sanitized for CR or LF characters when using `BlobBatchClient` APIs.
+
 ## 12.27.0-beta.1 (2026-07-22)
 
 ### Features Added